Create 7UP.cpp
There are N seats in a cricket stadium. As a part of the promotions for a match, every 7th seat gets a 7-UP T-Shirt. The first 200 seats are allocated for VIPs and last 400 seats are unreserved. These seats are not eligible for the T-shirts, Write a C++ program to find out which seats are eligible to get T-Shirt in the upcoming match.