Create Leaders.cpp
given an array A of positive integers. Find the leader in the order of appearance in the array.
Expected Time Complexity : 0(n)
Expected Auxiliary Space : 0(n)

Input: a[]={16,17,4,3,5,2}
output : 17 5 2 
Explanation: The first leader is 17 as it is greater than all the elements to its right. Similarly, the next leader is 5. The right most element is always a leader so it is also included.